Added connection module
