+++ backend/users/experimenter_factory.py
@@ -7,8 +7,8 @@ if TYPE_CHECKING:
 from aiortc import RTCSessionDescription
 
 from modules.filter_api import FilterAPI
-from modules.connection import connection_factory
-from modules.connection_subprocess import connection_subprocess_factory
+from connection.connection import connection_factory
+from connection.connection_subprocess import connection_subprocess_factory
 
 from users.experimenter import Experimenter
